refactor: 컴포넌트 폴더 내 ChatMainPage 폴더명 ChatPage로 변경, ChatDetailPage 폴더 삭제
